move helper functions from mongo planner code in separate file
